+++ src/internal/utils/ipfs/assertUploadContetResolve.ts
@@ -1,18 +1,14 @@
-import { HardhatPluginError } from 'hardhat/plugins'
+import { resolveRepoContentUri } from '../apm'
 
-/**
- * Sanity check to check if an IPFS API is active
- * Note: It requires the API to /api/v0/version route available
- */
-export async function assertIpfsApiIsAvailable(
-  ipfs: any,
-  url: string
+export async function assertUploadContetResolve(
+  contentHash: string,
+  gateway: string
 ): Promise<void> {
   try {
-    await ipfs.version()
-  } catch (e) {
-    throw new HardhatPluginError(
-      `IPFS API at ${url} is not available. Error: ${e}`
-    )
+    await resolveRepoContentUri(`ipfs:${contentHash}`, {
+      ipfsGateway: gateway,
+    })
+  } catch {
+    return
   }
 }
