+++ src/internal/utils/ipfs/index.ts
@@ -1,5 +1,5 @@
-export * from './assertIpfsApiIsAvailable'
 export * from './assertUploadContetResolve'
+export * from './assertIpfsApiIsAvailable'
 export * from './guessGatewayUrl'
 export * from './pinContent'
 export * from './uploadDirToIpfs'
